+++ FacilitEase/Data/AppDbContext.cs
@@ -1,0 +1,10 @@
+﻿using Microsoft.EntityFrameworkCore;
+namespace FacilitEase.Data
+{
+    public class AppDbContext : DbContext
+    {
+        public AppDbContext(DbContextOptions<AppDbContext> options) : base(options) { }
+
+
+    }
+}
