+++ FacilitEase/Repositories/IRepository.cs
@@ -1,10 +1,11 @@
-﻿using System;
-
-public interface IRepository<T>
+﻿namespace FacilitEase.Repositories
 {
-    IEnumerable<T> GetAll();
-    T GetById(int id);
-    void Add(T entity);
-    void Update(T entity);
-    void Delete(T entity);
+    public interface IRepository<T>
+    {
+        IEnumerable<T> GetAll();
+        T GetById(int id);
+        void Add(T entity);
+        void Update(T entity);
+        void Delete(T entity);
+    }
 }
