+++ FacilitEase/UnitOfWork/IUnitOfWork.cs
@@ -1,35 +1,8 @@
-﻿using System;
-
-namespace FacilitEase.UnitOfWork
+﻿namespace FacilitEase.UnitOfWork
 {
-    public interface UnitOfWork
+    public interface IUnitOfWork
     {
-        private readonly AppDbContext _context;
-
-        public UnitOfWork(AppDbContext context)
-        {
-            _context = context;
-            Employees = new EmployeeRepository(_context);
-            City = new CityRepository(_context);
-            Designation = new DesignationRepository(_context);
-
-            // Initialize other repositories.
-        }
-
-        public IEmployeeRepository Employees { get; }
-        public ICityRepository City { get; }
-
-        public IDesignationRepository Designation { get; }
-
-        public int Complete()
-        {
-            return _context.SaveChanges();
-        }
-
-        public void Dispose()
-        {
-            _context.Dispose();
-        }
-
+        int Complete();
+        void Dispose();
     }
 }
